requirements.txt
langchain
langgraph
faiss-cpu
openai
tiktoken
fastapi
uvicorn[standard]
python-dotenv
pydantic

